Perform PayU requests in separate controller
Previous method of making payment was involving a before_filter in
checkout controller, which was capturing and looking for
`payment_attributes`. This was very suboptimal.

The current approach is inspired by spree_paypal_express gem,
which renders a link to a separate controller, which handles the payment
api request and redirections.

I've just ported previous callback approach to an action-based approach.
I've also tried to implement Spree::PaymentMethod::PayU as a Gateway,
but given we need that separate controller the benefits were neglible.

Add gateway with translations

Remove any spree dependencies from Gemfile

Fix spec error related to comparing dates

Move callback to payu_controller, make first spec pass

Fix more specs

Fix all specs

Clean things up
