+++ Gemfile
@@ -4,8 +4,3 @@ source 'https://rubygems.org'
 # gem 'spree_auth_devise', github: 'spree/spree_auth_devise', branch: '3-0-stable'
 gem 'pry-rails'
 gemspec
-
-gem 'spree_frontend', path: '~/Devel/spree/frontend'
-gem 'spree_core', path: '~/Devel/spree/core'
-gem 'spree_backend', path: '~/Devel/spree/backend'
-gem 'spree_api', path: '~/Devel/spree/api'
