+++ app/views/spree/checkout/payment/_payu.html.erb
@@ -1,0 +1,1 @@
+<%= link_to Spree.t(:pay_with_payu), pay_with_payu_path(payment_method_id: payment_method.id) %>
